add MultisegmentWellEval
