+++ CMakeLists_files.cmake
@@ -63,6 +63,7 @@ list (APPEND MAIN_SOURCE_FILES
   opm/simulators/wells/GasLiftStage2.cpp
   opm/simulators/wells/GlobalWellInfo.cpp
   opm/simulators/wells/GroupState.cpp
+  opm/simulators/wells/MultisegmentWellEval.cpp
   opm/simulators/wells/MultisegmentWellGeneric.cpp
   opm/simulators/wells/ParallelWellInfo.cpp
   opm/simulators/wells/SegmentState.cpp
