+++ opm/simulators/wells/MultisegmentWell.hpp
@@ -23,7 +23,7 @@
 #define OPM_MULTISEGMENTWELL_HEADER_INCLUDED
 
 #include <opm/simulators/wells/WellInterface.hpp>
-#include <opm/simulators/wells/MultisegmentWellGeneric.hpp>
+#include <opm/simulators/wells/MultisegmentWellEval.hpp>
 
 #include <opm/parser/eclipse/EclipseState/Runspec.hpp>
 
@@ -33,10 +33,15 @@ namespace Opm
 
     template<typename TypeTag>
     class MultisegmentWell : public WellInterface<TypeTag>
-                           , public MultisegmentWellGeneric<GetPropType<TypeTag, Properties::Scalar>>
+                           , public MultisegmentWellEval<GetPropType<TypeTag, Properties::FluidSystem>,
+                                                         GetPropType<TypeTag, Properties::Indices>,
+                                                         GetPropType<TypeTag, Properties::Scalar>>
     {
     public:
-        typedef WellInterface<TypeTag> Base;
+        using Base = WellInterface<TypeTag>;
+        using MSWEval = MultisegmentWellEval<GetPropType<TypeTag, Properties::FluidSystem>,
+                                             GetPropType<TypeTag, Properties::Indices>,
+                                             GetPropType<TypeTag, Properties::Scalar>>;
 
         using typename Base::Simulator;
         using typename Base::IntensiveQuantities;
@@ -62,63 +67,21 @@ namespace Opm
         using Base::Oil;
         using Base::Gas;
 
-        // TODO: for now, not considering the polymer, solvent and so on to simplify the development process.
-
-        // TODO: we need to have order for the primary variables and also the order for the well equations.
-        // sometimes, they are similar, while sometimes, they can have very different forms.
-
-        // Table showing the primary variable indices, depending on what phases are present:
-        //
-        //         WOG     OG     WG     WO    W/O/G (single phase)
-        // GTotal    0      0      0      0                       0
-        // WFrac     1  -1000      1      1                   -1000
-        // GFrac     2      1  -1000  -1000                   -1000
-        // Spres     3      2      2      2                       1
-
-        static constexpr bool has_water = (Indices::waterSaturationIdx >= 0);
-        static constexpr bool has_gas = (Indices::compositionSwitchIdx >= 0);
-        static constexpr bool has_oil = (numPhases - has_gas - has_water) > 0;
-
-        // In the implementation, one should use has_wfrac_variable
-        // rather than has_water to check if you should do something
-        // with the variable at the WFrac location, similar for GFrac.
-        static constexpr bool has_wfrac_variable = has_water && numPhases > 1;
-        static constexpr bool has_gfrac_variable = has_gas && has_oil;
-
-        static constexpr int GTotal = 0;
-        static constexpr int WFrac = has_wfrac_variable ? 1 : -1000;
-        static constexpr int GFrac = has_gfrac_variable ? has_wfrac_variable + 1 : -1000;
-        static constexpr int SPres = has_wfrac_variable + has_gfrac_variable + 1;
-
-        //  the number of well equations  TODO: it should have a more general strategy for it
-        static const int numWellEq = numPhases + 1;
-
         using typename Base::Scalar;
 
         /// the matrix and vector types for the reservoir
         using typename Base::BVector;
         using typename Base::Eval;
 
-        // sparsity pattern for the matrices
-        // [A C^T    [x       =  [ res
-        //  B  D ]   x_well]      res_well]
-
-        // the vector type for the res_well and x_well
-        typedef Dune::FieldVector<Scalar, numWellEq> VectorBlockWellType;
-        typedef Dune::BlockVector<VectorBlockWellType> BVectorWell;
-
-        // the matrix type for the diagonal matrix D
-        typedef Dune::FieldMatrix<Scalar, numWellEq, numWellEq > DiagMatrixBlockWellType;
-        typedef Dune::BCRSMatrix <DiagMatrixBlockWellType> DiagMatWell;
-
-        // the matrix type for the non-diagonal matrix B and C^T
-        typedef Dune::FieldMatrix<Scalar, numWellEq, numEq>  OffDiagMatrixBlockWellType;
-        typedef Dune::BCRSMatrix<OffDiagMatrixBlockWellType> OffDiagMatWell;
-
-        // TODO: for more efficient implementation, we should have EvalReservoir, EvalWell, and EvalRerservoirAndWell
-        //                                                         EvalR (Eval), EvalW, EvalRW
-        // TODO: for now, we only use one type to save some implementation efforts, while improve later.
-        typedef DenseAd::Evaluation<double, /*size=*/numEq + numWellEq> EvalWell;
+        using typename MSWEval::EvalWell;
+        using typename MSWEval::BVectorWell;
+        using typename MSWEval::DiagMatWell;
+        using typename MSWEval::OffDiagMatrixBlockWellType;
+        using MSWEval::GFrac;
+        using MSWEval::WFrac;
+        using MSWEval::GTotal;
+        using MSWEval::SPres;
+        using MSWEval::numWellEq;
 
         MultisegmentWell(const Well& well,
                          const ParallelWellInfo& pw_info,
@@ -136,7 +99,6 @@ namespace Opm
                           const double gravity_arg,
                           const int num_cells,
                           const std::vector< Scalar >& B_avg) override;
-
 
         virtual void initPrimaryVariablesEvaluation() const override;
 
@@ -157,17 +119,15 @@ namespace Opm
                                        DeferredLogger& deferred_logger) const;
 
         /// check whether the well equations get converged for this well
-        virtual ConvergenceReport getWellConvergence(const WellState& well_state, const std::vector<double>& B_avg, DeferredLogger& deferred_logger, const bool relax_tolerance = false) const override;
+        virtual ConvergenceReport getWellConvergence(const WellState& well_state,
+                                                     const std::vector<double>& B_avg,
+                                                     DeferredLogger& deferred_logger,
+                                                     const bool relax_tolerance = false) const override;
 
         /// Ax = Ax - C D^-1 B x
         virtual void apply(const BVector& x, BVector& Ax) const override;
         /// r = r - C D^-1 Rw
         virtual void apply(BVector& r) const override;
-
-#if HAVE_CUDA || HAVE_OPENCL
-        /// add the contribution (C, D, B matrices) of this Well to the WellContributions object
-        void addWellContribution(WellContributions& wellContribs) const;
-#endif
 
         /// using the solution x to recover the solution xw for wells and applying
         /// xw to update Well State
@@ -260,66 +220,10 @@ namespace Opm
         using Base::calculateBhpFromThp;
         using Base::getALQ;
 
-        // TODO, the following should go to a class for computing purpose
-        // two off-diagonal matrices
-        mutable OffDiagMatWell duneB_;
-        mutable OffDiagMatWell duneC_;
-        // "diagonal" matrix for the well. It has offdiagonal entries for inlets and outlets.
-        mutable DiagMatWell duneD_;
-        /// \brief solver for diagonal matrix
-        ///
-        /// This is a shared_ptr as MultisegmentWell is copied in computeWellPotentials...
-        mutable std::shared_ptr<Dune::UMFPack<DiagMatWell> > duneDSolver_;
-
-        // residuals of the well equations
-        mutable BVectorWell resWell_;
-
-        // the values for the primary varibles
-        // based on different solutioin strategies, the wells can have different primary variables
-        mutable std::vector<std::array<double, numWellEq> > primary_variables_;
-
-        // the Evaluation for the well primary variables, which contain derivativles and are used in AD calculation
-        mutable std::vector<std::array<EvalWell, numWellEq> > primary_variables_evaluation_;
-
-        // depth difference between perforations and the perforated grid cells
-        std::vector<double> cell_perforation_depth_diffs_;
-        // pressure correction due to the different depth of the perforation and
-        // center depth of the grid block
-        std::vector<double> cell_perforation_pressure_diffs_;
-
         // the intial amount of fluids in each segment under surface condition
         std::vector<std::vector<double> > segment_fluid_initial_;
 
-        // the densities of segment fluids
-        // we should not have this member variable
-        std::vector<EvalWell> segment_densities_;
-
-        // the viscosity of the segments
-        std::vector<EvalWell> segment_viscosities_;
-
-        // the mass rate of the segments
-        std::vector<EvalWell> segment_mass_rates_;
-
-        // the upwinding segment for each segment based on the flow direction
-        std::vector<int> upwinding_segments_;
-
         mutable int debug_cost_counter_ = 0;
-
-        std::vector<std::vector<EvalWell>> segment_phase_fractions_;
-
-        std::vector<std::vector<EvalWell>> segment_phase_viscosities_;
-
-        std::vector<std::vector<EvalWell>> segment_phase_densities_;
-
-
-        void initMatrixAndVectors(const int num_cells) const;
-
-        EvalWell getBhp() const;
-        EvalWell getQs(const int comp_idx) const;
-        EvalWell getWQTotal() const;
-
-        // xw = inv(D)*(rw - C*x)
-        void recoverSolutionWell(const BVector& x, BVectorWell& xw) const;
 
         // updating the well_state based on well solution dwells
         void updateWellState(const BVectorWell& dwells,
@@ -334,16 +238,6 @@ namespace Opm
         // compute the pressure difference between the perforation and cell center
         void computePerfCellPressDiffs(const Simulator& ebosSimulator);
 
-        // fraction value of the primary variables
-        // should we just use member variables to store them instead of calculating them again and again
-        EvalWell volumeFraction(const int seg, const unsigned comp_idx) const;
-
-        // F_p / g_p, the basic usage of this value is because Q_p = G_t * F_p / G_p
-        EvalWell volumeFractionScaled(const int seg, const int comp_idx) const;
-
-        // basically Q_p / \sigma_p Q_p
-        EvalWell surfaceVolumeFraction(const int seg, const int comp_idx) const;
-
         void computePerfRatePressure(const IntensiveQuantities& int_quants,
                                      const std::vector<EvalWell>& mob_perfcells,
                                      const double Tw,
@@ -357,22 +251,9 @@ namespace Opm
                                      double& perf_vap_oil_rate,
                                      DeferredLogger& deferred_logger) const;
 
-        // convert a Eval from reservoir to contain the derivative related to wells
-        EvalWell extendEval(const Eval& in) const;
-
-        void updateThp(WellState& well_state, DeferredLogger& deferred_logger) const;
-
         // compute the fluid properties, such as densities, viscosities, and so on, in the segments
         // They will be treated implicitly, so they need to be of Evaluation type
         void computeSegmentFluidProperties(const Simulator& ebosSimulator);
-
-        EvalWell getSegmentPressure(const int seg) const;
-
-        EvalWell getSegmentRate(const int seg, const int comp_idx) const;
-
-        EvalWell getSegmentRateUpwinding(const int seg, const size_t comp_idx) const;
-
-        EvalWell getSegmentGTotal(const int seg) const;
 
         // get the mobility for specific perforation
         void getMobility(const Simulator& ebosSimulator,
@@ -391,32 +272,6 @@ namespace Opm
         std::vector<double>
         computeWellPotentialWithTHP(const Simulator& ebos_simulator,
                                     DeferredLogger& deferred_logger) const;
-
-        void assembleControlEq(const WellState& well_state,
-                               const GroupState& group_state,
-                               const Schedule& schedule,
-                               const SummaryState& summaryState,
-                               const Well::InjectionControls& inj_controls,
-                               const Well::ProductionControls& prod_controls,
-                               DeferredLogger& deferred_logger);
-
-        void assemblePressureEq(const int seg, const UnitSystem& unit_system,
-                                WellState& well_state, DeferredLogger& deferred_logger) const;
-
-        void assembleDefaultPressureEq(const int seg, WellState& well_state) const;
-
-        // hytrostatic pressure loss
-        EvalWell getHydroPressureLoss(const int seg) const;
-
-        // frictinal pressure loss
-        EvalWell getFrictionPressureLoss(const int seg) const;
-
-        void handleAccelerationPressureLoss(const int seg, WellState& well_state) const;
-
-        // handling the overshooting and undershooting of the fractions
-        void processFractions(const int seg) const;
-
-        void updateWellStateFromPrimaryVariables(WellState& well_state, DeferredLogger& deferred_logger) const;
 
         virtual double getRefDensity() const override;
 
@@ -440,21 +295,6 @@ namespace Opm
 
         EvalWell getSegmentSurfaceVolume(const Simulator& ebos_simulator, const int seg_idx) const;
 
-        std::vector<Scalar> getWellResiduals(const std::vector<Scalar>& B_avg,
-                                             DeferredLogger& deferred_logger) const;
-
-        double getResidualMeasureValue(const WellState& well_state,
-                                       const std::vector<double>& residuals,
-                                       DeferredLogger& deferred_logger) const;
-
-        double getControlTolerance(const WellState& well_state, DeferredLogger& deferred_logger) const;
-
-        void checkConvergenceControlEq(const WellState& well_state,
-                                       ConvergenceReport& report,
-                                       DeferredLogger& deferred_logger) const;
-
-        void updateUpwindingSegments();
-
         // turn on crossflow to avoid singular well equations
         // when the well is banned from cross-flow and the BHP is not properly initialized,
         // we turn on crossflow to avoid singular well equations. It can result in wrong-signed
@@ -477,19 +317,6 @@ namespace Opm
 
         double maxPerfPress(const Simulator& ebos_simulator) const;
 
-        // pressure drop for Spiral ICD segment (WSEGSICD)
-        EvalWell pressureDropSpiralICD(const int seg) const;
-
-        // pressure drop for Autonomous ICD segment (WSEGAICD)
-        EvalWell pressureDropAutoICD(const int seg, const UnitSystem& unit_system) const;
-
-        // pressure drop for sub-critical valve (WSEGVALV)
-        EvalWell pressureDropValve(const int seg) const;
-
-        // assemble pressure equation for ICD segments
-        void assembleICDPressureEq(const int seg, const UnitSystem& unit_system,
-                                   WellState& well_state, DeferredLogger& deferred_logger) const;
-
         // check whether the well is operable under BHP limit with current reservoir condition
         virtual void checkOperabilityUnderBHPLimitProducer(const WellState& well_state, const Simulator& ebos_simulator, DeferredLogger& deferred_logger) override;
 
@@ -498,7 +325,6 @@ namespace Opm
 
         // updating the inflow based on the current reservoir condition
         virtual void updateIPR(const Simulator& ebos_simulator, DeferredLogger& deferred_logger) const override;
-
     };
 
 }
